[REF] multiple_worksheets: moved meta to separate file, added comments
