+++ spence_project_multiple_worksheets/report/worksheet_custom_reports.xml
@@ -13,7 +13,7 @@
         </record>
 
         <record id="mail_template_data_send_report" model="mail.template">
-            <field name="name">Task Report</field>
+            <field name="name">Worksheet Report</field>
             <field name="model_id" ref="spence_project_multiple_worksheets.model_project_task_worksheet_meta"/>
             <field name="subject">${object.name} Report</field>
             <field name="email_to">${(object.partner_id.email_formatted) |safe}</field>
